+++ src/utilities.jl
@@ -96,31 +96,44 @@ function rownormalize(A::LowerTriangular{T, Diagonal{T, Vector{T}}}) where T
     one(T) * I(size(A,1))
 end
 
+# from the ProgressMeter docs
+_is_logging(io) = isa(io, Base.TTY) == false || (get(ENV, "CI", nothing) == "true")
+
 """
     replicate(f::Function, n::Integer; use_threads=false)
 
 Return a vector of the values of `n` calls to `f()` - used in simulations where the value of `f` is stochastic.
 
-Note that if `f()` is not thread-safe or depends on a non thread-safe RNG,
+`hide_progress` can be used to disable the progress bar. Note that the progress
+bar is automatically disabled for non-interactive (i.e. logging) contexts.
+
+!!! warning
+    If `f()` is not thread-safe or depends on a non thread-safe RNG,
     then you must set `use_threads=false`. Also note that ordering of replications
     is not guaranteed when `use_threads=true`, although the replications are not
     otherwise affected for thread-safe `f()`.
 """
-function replicate(f::Function, n::Integer; use_threads=false)
+function replicate(f::Function, n::Integer;
+                   use_threads=false, hide_progress=false)
+    # no macro version yet: https://github.com/timholy/ProgressMeter.jl/issues/143
+    # and we want some advanced options
+    p = Progress(n; output=Base.stderr,
+                 enabled=!hide_progress && !_is_logging(stderr))
+    # get the type
+    rr = f()
+    next!(p)
+    # pre-allocate
+    results = [rr for _ in Base.OneTo(n)]
     if use_threads
-        # no macro version yet: https://github.com/timholy/ProgressMeter.jl/issues/143
-        p = Progress(n)
-        # get the type
-        rr = f()
-        next!(p)
-        # pre-allocate
-        results = [rr for _ in Base.OneTo(n)]
         Threads.@threads for idx = 2:n
             results[idx] = f()
             next!(p)
         end
     else
-        results = @showprogress [f() for _ in Base.OneTo(n)]
+        for idx = 2:n
+            results[idx] = f()
+            next!(p)
+        end
     end
     results
 end